requirements.txt
altair
pandas
numpy
rasterio
jupyterlab
jupyter-book
matplotlib
tqdm
folium
leafmap
uvicorn 
titiler.application
rangehttpserver
streamlit_folium